move OnednnGraph helper class to the onednn_graph_fusion.py
